# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Riad Lumière Luxury Hotel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Room {
  id: number;
  name: string;
  /** riad-suite | grand-suite | deluxe-room | prestige-room | royal-pavilion */
  category: string;
  description: string;
  shortDescription?: string;
  pricePerNight: number;
  maxGuests: number;
  /** Square meters */
  size: number;
  bedType?: string;
  floor?: number;
  images: string[];
  amenities: string[];
  available: boolean;
  featured?: boolean;
}

export interface RoomInput {
  name: string;
  category: string;
  description: string;
  shortDescription?: string;
  pricePerNight: number;
  maxGuests: number;
  size: number;
  bedType?: string;
  floor?: number;
  images?: string[];
  amenities?: string[];
  featured?: boolean;
}

export interface RoomUpdate {
  name?: string;
  category?: string;
  description?: string;
  pricePerNight?: number;
  maxGuests?: number;
  available?: boolean;
  featured?: boolean;
}

export interface RoomWithStatus {
  id: number;
  name: string;
  category: string;
  pricePerNight: number;
  available: boolean;
  /** @nullable */
  currentBookingId: number | null;
  /** @nullable */
  nextCheckIn?: string | null;
  /** @nullable */
  nextCheckOut?: string | null;
}

export interface AvailabilityResult {
  checkIn: string;
  checkOut: string;
  guests: number;
  nights?: number;
  availableRooms: Room[];
}

export interface Booking {
  id: number;
  roomId: number;
  roomName?: string;
  checkIn: string;
  checkOut: string;
  guests: number;
  nights?: number;
  /** pending | confirmed | cancelled | completed */
  status: string;
  totalPrice: number;
  guestName?: string;
  guestEmail?: string;
  guestPhone?: string;
  specialRequests?: string;
  confirmationCode: string;
  createdAt?: string;
}

export interface BookingInput {
  roomId: number;
  checkIn: string;
  checkOut: string;
  guests: number;
  guestName: string;
  guestEmail: string;
  guestPhone?: string;
  specialRequests?: string;
}

export interface BookingUpdate {
  status?: string;
  specialRequests?: string;
  guestPhone?: string;
}

export interface PaymentInput {
  cardholderName: string;
  cardLast4: string;
  paymentMethod: string;
}

export interface Experience {
  id: number;
  name: string;
  /** spa | wellness | tour | cultural | dining | concierge */
  category: string;
  description: string;
  shortDescription?: string;
  duration: string;
  price: number;
  maxParticipants?: number;
  images: string[];
  featured?: boolean;
  included?: string[];
}

export interface Restaurant {
  id: number;
  name: string;
  cuisine: string;
  description: string;
  shortDescription?: string;
  dressCode?: string;
  openHours?: string;
  capacity?: number;
  chefName?: string;
  chefBio?: string;
  images: string[];
  menuHighlights?: string[];
  featured?: boolean;
}

export interface DiningReservationInput {
  date: string;
  time: string;
  guests: number;
  guestName: string;
  guestEmail: string;
  specialRequests?: string;
}

export interface DiningReservation {
  id: number;
  restaurantId: number;
  restaurantName?: string;
  date: string;
  time: string;
  guests: number;
  guestName: string;
  guestEmail: string;
  specialRequests?: string;
  confirmationCode: string;
}

export interface GalleryItem {
  id: number;
  url: string;
  /** photo | video */
  type: string;
  /** rooms | spa | dining | garden | architecture | lifestyle */
  category: string;
  alt: string;
  caption?: string;
  featured?: boolean;
}

export interface Review {
  id: number;
  guestName: string;
  country: string;
  rating: number;
  title: string;
  body: string;
  stayDate: string;
  roomName?: string;
  featured: boolean;
  avatarUrl?: string;
}

export interface ReviewInput {
  guestName: string;
  country: string;
  rating: number;
  title: string;
  body: string;
  stayDate: string;
  roomName?: string;
}

export interface NewsletterInput {
  email: string;
  firstName?: string;
}

export interface NewsletterSubscription {
  id: number;
  email: string;
  firstName?: string;
  subscribedAt: string;
}

export interface ContactInput {
  name: string;
  email: string;
  phone?: string;
  subject: string;
  message: string;
  preferredContact?: string;
}

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  phone?: string;
  subject: string;
  message: string;
  createdAt: string;
}

export interface AdminStats {
  totalBookings: number;
  activeBookings: number;
  totalRevenue: number;
  occupancyRate: number;
  avgNightlyRate: number;
  avgStayLength: number;
  totalRooms: number;
  totalGuests: number;
  revenueThisMonth?: number;
  bookingsThisMonth?: number;
}

export interface MonthlyOccupancy {
  month: string;
  year: number;
  occupancyRate: number;
  bookedNights: number;
  totalNights: number;
}

export interface MonthlyRevenue {
  month: string;
  year: number;
  revenue: number;
  bookings: number;
}

export type ListRoomsParams = {
category?: string;
maxGuests?: number;
};

export type CheckAvailabilityParams = {
checkIn: string;
checkOut: string;
guests: number;
rooms?: number;
};

export type ListBookingsParams = {
status?: string;
limit?: number;
offset?: number;
};

export type ListExperiencesParams = {
category?: string;
};

export type ListGalleryItemsParams = {
category?: string;
type?: string;
};

export type ListReviewsParams = {
featured?: boolean;
limit?: number;
};

export type GetOccupancyByMonthParams = {
year?: number;
};

export type GetRevenueByMonthParams = {
year?: number;
};