# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Riad Lumière Luxury Hotel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    delete:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/confirm:
    post:
      operationId: confirmBooking
      tags: [bookings]
      summary: Confirm and finalize booking (payment step)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # --- EXPERIENCES ---
  /experiences:
    get:
      operationId: listExperiences
      tags: [experiences]
      summary: List all experiences
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of experiences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Experience"
  /experiences/{id}:
    get:
      operationId: getExperience
      tags: [experiences]
      summary: Get experience details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Experience details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"

  # --- RESTAURANTS ---
...
  /admin/occupancy:
    get:
      operationId: getOccupancyByMonth
      tags: [admin]
      summary: Get occupancy breakdown by month
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly occupancy data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyOccupancy"
  /admin/revenue:
    get:
      operationId: getRevenueByMonth
      tags: [admin]
      summary: Get revenue breakdown by month
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly revenue data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyRevenue"
  /admin/rooms:
    get:
      operationId: listAdminRooms
      tags: [admin]
      summary: Admin room list with status
      responses:
        "200":
          description: Rooms with status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoomWithStatus"
    post:
      operationId: createRoom
      tags: [admin]
      summary: Create a new room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /admin/rooms/{id}:
    patch:
      operationId: updateRoom
      tags: [admin]
      summary: Update room details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomUpdate"
      responses:
        "200":
          description: Updated room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      operationId: deleteRoom
      tags: [admin]
      summary: Delete a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
components:
  schemas:
...
    Room:
      type: object
      required: [id, name, category, description, pricePerNight, maxGuests, size, images, amenities, available]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          description: "riad-suite | grand-suite | deluxe-room | prestige-room | royal-pavilion"
        description:
          type: string
        shortDescription:
          type: string
        pricePerNight:
          type: number
        maxGuests:
          type: integer
        size:
          type: integer
          description: Square meters
        bedType:
          type: string
        floor:
          type: integer
        images:
          type: array
          items:
            type: string
        amenities:
          type: array
          items:
            type: string
        available:
          type: boolean
        featured:
          type: boolean
    RoomInput:
      type: object
      required: [name, category, description, pricePerNight, maxGuests, size]
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        pricePerNight:
          type: number
        maxGuests:
          type: integer
        size:
          type: integer
        bedType:
          type: string
        floor:
          type: integer
        images:
          type: array
          items:
            type: string
        amenities:
          type: array
          items:
            type: string
        featured:
          type: boolean
    RoomUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        pricePerNight:
          type: number
        maxGuests:
          type: integer
        available:
          type: boolean
        featured:
          type: boolean
    RoomWithStatus:
      type: object
      required: [id, name, category, pricePerNight, available, currentBookingId]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        pricePerNight:
          type: number
        available:
          type: boolean
        currentBookingId:
          type: ["integer", "null"]
        nextCheckIn:
          type: ["string", "null"]
        nextCheckOut:
          type: ["string", "null"]
...
    Booking:
      type: object
      required: [id, roomId, checkIn, checkOut, guests, status, totalPrice, confirmationCode]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        roomName:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        guests:
          type: integer
        nights:
          type: integer
        status:
          type: string
          description: "pending | confirmed | cancelled | completed"
        totalPrice:
          type: number
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        specialRequests:
          type: string
        confirmationCode:
          type: string
        createdAt:
          type: string
    BookingInput:
      type: object
      required: [roomId, checkIn, checkOut, guests, guestName, guestEmail]
      properties:
        roomId:
          type: integer
        checkIn:
          type: string
        checkOut:
          type: string
        guests:
          type: integer
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        specialRequests:
          type: string
    BookingUpdate:
      type: object
      properties:
        status:
          type: string
        specialRequests:
          type: string
        guestPhone:
          type: string
    PaymentInput:
      type: object
      required: [cardholderName, cardLast4, paymentMethod]
      properties:
        cardholderName:
          type: string
        cardLast4:
          type: string
        paymentMethod:
          type: string
    Experience:
      type: object
      required: [id, name, category, description, duration, price, images]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          description: "spa | wellness | tour | cultural | dining | concierge"
        description:
          type: string
        shortDescription:
          type: string
        duration:
          type: string
        price:
          type: number
        maxParticipants:
          type: integer
        images:
          type: array
          items:
            type: string
        featured:
          type: boolean
        included:
          type: array
          items:
            type: string
...
    MonthlyOccupancy:
      type: object
      required: [month, year, occupancyRate, bookedNights, totalNights]
      properties:
        month:
          type: string
        year:
          type: integer
        occupancyRate:
          type: number
        bookedNights:
          type: integer
        totalNights:
          type: integer
    MonthlyRevenue:
      type: object
      required: [month, year, revenue, bookings]
      properties:
        month:
          type: string
        year:
          type: integer
        revenue:
          type: number
        bookings:
          type: integer