# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Riad Lumière Luxury Hotel API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  roomId: number;
  roomName?: string;
  checkIn: string;
  checkOut: string;
  guests: number;
  nights?: number;
  /** pending | confirmed | cancelled | completed */
  status: string;
  totalPrice: number;
  guestName?: string;
  guestEmail?: string;
  guestPhone?: string;
  specialRequests?: string;
  confirmationCode: string;
  createdAt?: string;
}