# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Riad Lumière Luxury Hotel API
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: number;
  name: string;
  /** riad-suite | grand-suite | deluxe-room | prestige-room | royal-pavilion */
  category: string;
  description: string;
  shortDescription?: string;
  pricePerNight: number;
  maxGuests: number;
  /** Square meters */
  size: number;
  bedType?: string;
  floor?: number;
  images: string[];
  amenities: string[];
  available: boolean;
  featured?: boolean;
}